api/requirements.txt
fastapi
uvicorn
Pillow
keras
tensorflow
numpy
